Practice string

+++ newc/factre.c
@@ -1,15 +1,15 @@
 #include<stdio.h>
 
-long int multiplyNumbers(int n);
+unsigned long long multiplyNumbers(int n);
 int main() {
     int n;
     printf("Enter a positive integer: ");
     scanf("%d",&n);
-    printf("Factorial of %d = %ld\n", n, multiplyNumbers(n));
+    printf("Factorial of %d = %llu\n", n, multiplyNumbers(n));
     return 0;
 }
 
-long int multiplyNumbers(int n) {
+unsigned long long multiplyNumbers(int n) {
     if (n>=1)
         return n*multiplyNumbers(n-1);
     else
